refactor(components): Input prefix 삭제
